Outsource the route configuration
Use the app-routing.module file to handle all route configuration in the app.
