+++ src/app/app-routing.module.ts
@@ -1,0 +1,35 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { PageNotFoundComponent } from 'app/page-not-found/page-not-found.component';
+import { EditServerComponent } from 'app/servers/edit-server/edit-server.component';
+import { ServerComponent } from 'app/servers/server/server.component';
+import { ServersComponent } from 'app/servers/servers.component';
+import { UserComponent } from 'app/users/user/user.component';
+import { UsersComponent } from 'app/users/users.component';
+import { HomeComponent } from 'app/home/home.component';
+
+const appRoutes: Routes = [
+  { path: '', component: HomeComponent },
+  {
+    path: 'users', component: UsersComponent, children: [
+      { path: ':id/:name', component: UserComponent }
+    ]
+  },
+  {
+    path: 'servers', component: ServersComponent, children: [
+      { path: ':id', component: ServerComponent },
+      { path: ':id/edit', component: EditServerComponent }
+    ]
+  },
+  { path: 'not-found', component: PageNotFoundComponent },
+  { path: '**', redirectTo: '/not-found' }
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forRoot(appRoutes)
+  ],
+  exports: [RouterModule]
+})
+export class AppRoutingModule { }
